docs: fully documented
