+++ @dekk/countdown/src/date-helpers.js
@@ -1,0 +1,20 @@
+/**
+ * Helpers to calculate time units into milliseconds
+ * @private
+ * @property {number} s
+ *   Use for seconds
+ * @property {number} m
+ *   Use for minutes
+ * @property {number} h
+ *   Use for hours
+ * @property {number} d
+ *   Use for days
+ */
+const dateHelpers = {
+  s: 1000,
+  m: 60000,
+  h: 3600000,
+  d: 86400000
+}
+
+export default dateHelpers
