+++ @dekk/countdown/src/index.js
@@ -1,17 +1,28 @@
 /* global window */
 import {Component} from 'react'
 import PropTypes from 'prop-types'
+import dateHelpers from './date-helpers'
 
 export {default as renderCountdown} from './render-countdown'
 
-const dateHelpers = {
-  s: 1000,
-  m: 60000,
-  h: 3600000,
-  d: 86400000
-}
-
+/**
+ * A countdowntimer that can be paused.
+ * @public
+ * @param {Object} props
+ *   The properties
+ * @param {Boolean} props.isRunning
+ *   Defines if the timer is running or paused
+ * @param {number} props.timerWarning
+ *   Defines when the timer warns you
+ * @param {number} props.end
+ *   Defines when the timerends
+ * @param {renderCountdown} props.render
+ *   A function that can be passed to render the timer
+ */
 export default class Countdown extends Component {
+  /**
+   * @private
+   */
   static get propTypes() {
     return {
       isRunning: PropTypes.bool,
@@ -21,6 +32,9 @@ export default class Countdown extends Component {
     }
   }
 
+  /**
+   * @private
+   */
   static get defaultProps() {
     return {
       isRunning: false,
@@ -28,31 +42,66 @@ export default class Countdown extends Component {
     }
   }
 
-  state = {
-    days: 0,
-    hours: 0,
-    minutes: 0,
-    seconds: 0,
-    start: this.now.getTime(),
-    waiting: this.props.isRunning ? 0 : this.now.getTime(),
-    waited: 0
-  }
-
+  /**
+   * A countdowntimer that can be paused.
+   * @public
+   * @param {Object} props
+   *   The properties
+   * @param {Boolean} props.isRunning
+   *   Defines if the timer is running or paused
+   * @param {number} props.timerWarning
+   *   Defines when the timer warns you
+   * @param {number} props.end
+   *   Defines when the timerends
+   * @param {renderCountdown} props.render
+   *   A function that can be passed to render the timer
+   */
   constructor(props) {
     super(props)
+
+    /**
+     * @private
+     * @property {number} days
+     * @property {number} hours
+     * @property {number} minutes
+     * @property {number} seconds
+     * @property {number} start
+     * @property {number} waiting
+     * @property {number} waited
+     */
+    this.state = {
+      days: 0,
+      hours: 0,
+      minutes: 0,
+      seconds: 0,
+      start: this.now.getTime(),
+      waiting: this.props.isRunning ? 0 : this.now.getTime(),
+      waited: 0
+    }
 
     this.run = this.run.bind(this)
     this.stop = this.stop.bind(this)
   }
 
+  /**
+   * @private
+   */
   get now() {
     return new Date()
   }
 
+  /**
+   * @private
+   */
   get then() {
     return this.state.start + this.state.waited + this.props.end * dateHelpers.m
   }
 
+  /**
+   * @private
+   * @param {(Boolean|number)} force
+   *   If a `Boolean` use it to force runs.
+   */
   run(force = false) {
     const now = this.now
     const then = this.then
@@ -76,10 +125,16 @@ export default class Countdown extends Component {
     }
   }
 
+  /**
+   * @private
+   */
   stop() {
     window.cancelAnimationFrame(this.run)
   }
 
+  /**
+   * @private
+   */
   componentWillMount() {
     const now = this.now.getTime()
     const then = this.state.waited + this.props.end * dateHelpers.m
@@ -100,6 +155,10 @@ export default class Countdown extends Component {
     }
   }
 
+  /**
+   * @private
+   * @param {Object} newProps
+   */
   componentWillReceiveProps(newProps) {
     if (newProps.isRunning && !this.props.isRunning) {
       const {waiting} = this.state
@@ -115,10 +174,16 @@ export default class Countdown extends Component {
     }
   }
 
+  /**
+   * @private
+   */
   componentWillUnmount() {
     this.stop()
   }
 
+  /**
+   * @private
+   */
   render() {
     const {days, hours, minutes, seconds, done} = this.state
     return this.props.render(
@@ -133,3 +198,23 @@ export default class Countdown extends Component {
     )
   }
 }
+
+/**
+ * @typedef renderCountdown
+ * @public
+ * @param {Object} data
+ *   Time data
+ * @param {number} data.days
+ *   Remaining days
+ * @param {number} data.hours
+ *   Remaining hours
+ * @param {number} data.minutes
+ *   Remaining minutes
+ * @param {number} data.seconds
+ *   Remaining seconds
+ * @param {Boolean} isDone
+ *   Triggered when the timer is done
+ * @param {Boolean} timerWarning
+ *   Triggered when the warning limit has been reached
+ * @return {ReactElement}
+ */
