+++ @dekk/deck/src/notes.js
@@ -2,6 +2,12 @@ import React, {Children, cloneElement} from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
+/**
+ * A Note provider to help with store based notes
+ * @private
+ * @param {Object} props
+ * @param {(ReactElement|ReactElement[])} props.children
+ */
 const NoteProvider = ({children}) => (
   <div>
     {Children.toArray(children).map((child, i) =>
@@ -10,16 +16,26 @@ const NoteProvider = ({children}) => (
   </div>
 )
 
+/**
+ * @private
+ */
 NoteProvider.propTypes = {
   children: PropTypes.node
 }
 
+/**
+ * @private
+ */
 NoteProvider.defaultProps = {
   children: null
 }
 
 export default NoteProvider
 
+/**
+ * A Notes section for the speaker
+ * @private
+ */
 export const Notes = styled.div`
   position: relative;
   height: 100%;
