+++ @dekk/deck/src/views.js
@@ -1,6 +1,9 @@
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
+/**
+ * @private
+ */
 const View = styled.div`
   --scale: var(--view-scale);
   position: relative;
@@ -14,10 +17,16 @@ const View = styled.div`
 
 export default View
 
+/**
+ * @private
+ */
 export const Preload = styled.div`
   display: none;
 `
 
+/**
+ * @private
+ */
 export const Preview = styled.div`
   --scale: var(--preview-scale);
   position: relative;
@@ -29,6 +38,9 @@ export const Preview = styled.div`
   overflow: hidden;
 `
 
+/**
+ * @private
+ */
 export const Nextview = styled.div`
   ${({layout}) => (layout === 0 ? '' : 'display: none')};
   --scale: var(--nextview-scale);
@@ -41,10 +53,16 @@ export const Nextview = styled.div`
   overflow: hidden;
 `
 
+/**
+ * @private
+ */
 Nextview.propTypes = {
   layout: PropTypes.number
 }
 
+/**
+ * @private
+ */
 Nextview.defaultProps = {
   layout: 0
 }
